selects a seeker at random from the players. Fixed a bug where players were deleted after game starts

+++ campus_game_project/game/consumers.py
@@ -2,6 +2,7 @@ import json
 from asgiref.sync import async_to_sync
 from channels.generic.websocket import WebsocketConsumer
 from game.models import Game, Player
+from random import randint
 
 
 class PlayerConsumer(WebsocketConsumer):
@@ -43,23 +44,27 @@ class PlayerConsumer(WebsocketConsumer):
 
     # Behaviour when the user disconnects
     def disconnect(self, close_code):
-        message = self.scope['session']['username'] + " has left."
 
-        # Sends data to group
-        async_to_sync(self.channel_layer.group_send)(
-            self.lobby_code,
-            {
-                'type': 'lobby_event',
-                'msg_type': 'leave',
-                'message': message,
-                'username': self.scope['session']['username'],
-                'players': None
-            }
-        )
+        if "playing" not in self.scope['session']:
+            message = self.scope['session']['username'] + " has left."
 
-        # Deletes the user from the database
-        g = Game.objects.get(lobby_code=self.lobby_code)
-        Player.objects.get(game=g, username=self.scope['session']['username']).delete()
+            # Sends data to group
+            async_to_sync(self.channel_layer.group_send)(
+                self.lobby_code,
+                {
+                    'type': 'lobby_event',
+                    'msg_type': 'leave',
+                    'message': message,
+                    'username': self.scope['session']['username'],
+                    'players': None
+                }
+            )
+
+            # Deletes the user from the database
+            g = Game.objects.get(lobby_code=self.lobby_code)
+            Player.objects.get(game=g, username=self.scope['session']['username']).delete()
+            g.player_num -= 1
+            g.save()
 
         # Leaves the group
         async_to_sync(self.channel_layer.group_discard)(
@@ -70,39 +75,47 @@ class PlayerConsumer(WebsocketConsumer):
     # Behaviour when the websocket receives a message
     def receive(self, text_data):
         text_data_json = json.loads(text_data)
-        ready = text_data_json['ready'].lower()
-        username = text_data_json['username']
-        message = username + " is " + ready + "."
+        if text_data_json['msg_type'] == "playing":
+            self.scope['session']['playing'] = True
+        elif text_data_json['msg_type'] == "ready":
+            ready = text_data_json['ready'].lower()
+            username = text_data_json['username']
+            message = username + " is " + ready + "."
 
-        if ready == "unready":
-            ready = False
-        else:
-            ready = True
+            if ready == "unready":
+                ready = False
+            else:
+                ready = True
 
-        # Sets the player to ready
-        g = Game.objects.get(lobby_code=self.lobby_code)
-        p = Player.objects.get(game=g, username=username)
-        p.ready = ready
-        p.save()
+            # Sets the player to ready
+            g = Game.objects.get(lobby_code=self.lobby_code)
+            p = Player.objects.get(game=g, username=username)
+            p.ready = ready
+            p.save()
 
-        if g.all_ready():
+            if g.all_ready():
+                players = Player.objects.filter(game=g)
+                p = players[randint(0, len(players)-1)]
+                p.seeker = True
+                p.save()
+
+                async_to_sync(self.channel_layer.group_send)(
+                    self.lobby_code,
+                    {
+                        'type': 'start_game'
+                    }
+                )
+
+            # Sends message to group
             async_to_sync(self.channel_layer.group_send)(
                 self.lobby_code,
                 {
-                    'type': 'start_game'
+                    'type': 'ready_event',
+                    'message': message,
+                    'username': self.scope['session']['username'],
+                    'ready': ready
                 }
             )
-
-        # Sends message to group
-        async_to_sync(self.channel_layer.group_send)(
-            self.lobby_code,
-            {
-                'type': 'ready_event',
-                'message': message,
-                'username': self.scope['session']['username'],
-                'ready': ready
-            }
-        )
 
     # A user joins or leaves
     def lobby_event(self, event):
